KT-19943: Remove redundant type conversions in J2kPostProcessing
Don't remove them in converter itself

+++ j2k/testData/fileOrElement/issues/kt-851.kt
@@ -3,6 +3,6 @@ internal class Test {
 
     fun test() {
         val b = 10
-        putInt(b.toInt())
+        putInt(b)
     }
 }
